move np home into a component
